Encapsulate the socket in an ADT.
Create an envelope-like ADT and store the socket's file descriptor in
the ADT instead of using it directly.

This is step one in the process of adding SSL support to meta_socket and
tcp_server.

+++ src/meta_stack.c
@@ -122,7 +122,7 @@ void* stack_get(stack s, size_t i)
 
 int main(void)
 {
-	int i, nelem = 10000;
+	size_t i, nelem = 10000;
 	char *str;
 	stack s = stack_new();
 
@@ -132,7 +132,7 @@ int main(void)
 			exit(EXIT_FAILURE);
 		}
 
-		sprintf(str, "%d", i);
+		sprintf(str, "%zu", i);
 		stack_push(s, str);
 	}
